ai/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
python-dotenv>=1.0
torch>=2.2
transformers>=4.40
accelerate>=0.30
requests>=2.31
peft>=0.11
datasets>=2.19
safetensors>=0.4
bitsandbytes>=0.43; platform_system == "Linux"
